requirements.txt
backcall==0.2.0
colorama==0.4.4
decorator==4.4.2
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.16
Pygments==2.8.0
python-dotenv==0.15.0
scapy==2.4.4
traitlets==5.0.5
wcwidth==0.2.5
